// app/schemas/auth.ts
import z from "zod";

export const signUpSchema = z.object({
    name: z.string().min(3, "Name must be at least 3 characters long").max(18, "Name must be at most 18 characters long"),
    email: z.email(),
    password: z.string().min(8, "Password must be at least 8 characters long").max(50, "Password must be at most 50 characters long")
})

export const loginSchema = z.object({
    email: z.email(),
    password: z.string().min(8, "Password must be at least 8 characters long").max(50, "Password must be at most 50 characters long")
})